modify: header file dependencies
