+++ Release/src/common/algorithm.d
@@ -1,12 +1,9 @@
 src/common/algorithm.o: ../src/common/algorithm.cpp \
- ../src/common/./algorithm.hpp ../src/common/./../typing/types.hpp \
- ../src/common/./../typing/./stl_wrapper/weak_ptr.hpp \
- ../src/common/./_v_func.hpp
+ ../src/common/./algorithm.hpp ../src/common/./../typing/gtypes/types.hpp \
+ ../src/common/./../typing/gtypes/./stl_wrapper/weak_ptr.hpp
 
 ../src/common/./algorithm.hpp:
 
-../src/common/./../typing/types.hpp:
+../src/common/./../typing/gtypes/types.hpp:
 
-../src/common/./../typing/./stl_wrapper/weak_ptr.hpp:
-
-../src/common/./_v_func.hpp:
+../src/common/./../typing/gtypes/./stl_wrapper/weak_ptr.hpp:
