+++ Release/src/common/vec1d.d
@@ -1,34 +1,38 @@
 src/common/vec1d.o: ../src/common/vec1d.cpp ../src/common/./vec1d.hpp \
- ../src/common/././algorithm.cpp ../src/common/./././algorithm.hpp \
- ../src/common/./././../typing/types.hpp \
- ../src/common/./././../typing/./stl_wrapper/weak_ptr.hpp \
- ../src/common/./././_v_func.hpp ../src/common/./../typing/allocator.cpp \
+ ../src/common/./../typing/gtypes/types.hpp \
+ ../src/common/./../typing/gtypes/./stl_wrapper/weak_ptr.hpp \
+ ../src/common/./../handlers/exception.hpp \
+ ../src/common/./../typing/allocator.cpp \
  ../src/common/./../typing/./allocator.hpp \
- ../src/common/./../typing/././types.hpp \
+ ../src/common/./../typing/././gtypes/types.hpp \
  ../src/common/./../random/generator.cpp \
  ../src/common/./../random/./generator.hpp \
- ../src/common/./../handlers/exception.hpp
+ ../src/common/././vec1d_ufunc.hpp ../src/common/././algorithm.cpp \
+ ../src/common/./././algorithm.hpp \
+ ../src/common/./././../typing/gtypes/types.hpp
 
 ../src/common/./vec1d.hpp:
 
-../src/common/././algorithm.cpp:
+../src/common/./../typing/gtypes/types.hpp:
 
-../src/common/./././algorithm.hpp:
+../src/common/./../typing/gtypes/./stl_wrapper/weak_ptr.hpp:
 
-../src/common/./././../typing/types.hpp:
-
-../src/common/./././../typing/./stl_wrapper/weak_ptr.hpp:
-
-../src/common/./././_v_func.hpp:
+../src/common/./../handlers/exception.hpp:
 
 ../src/common/./../typing/allocator.cpp:
 
 ../src/common/./../typing/./allocator.hpp:
 
-../src/common/./../typing/././types.hpp:
+../src/common/./../typing/././gtypes/types.hpp:
 
 ../src/common/./../random/generator.cpp:
 
 ../src/common/./../random/./generator.hpp:
 
-../src/common/./../handlers/exception.hpp:
+../src/common/././vec1d_ufunc.hpp:
+
+../src/common/././algorithm.cpp:
+
+../src/common/./././algorithm.hpp:
+
+../src/common/./././../typing/gtypes/types.hpp:
