+++ Release/src/iterators/SequentialNdIterator.d
@@ -1,47 +1,56 @@
 src/iterators/SequentialNdIterator.o: \
  ../src/iterators/SequentialNdIterator.cpp \
  ../src/iterators/SequentialNdIterator.hpp \
- ../src/iterators/../shapes/coords.hpp \
- ../src/iterators/../shapes/../common/vec1d.cpp \
- ../src/iterators/../shapes/../common/./vec1d.hpp \
- ../src/iterators/../shapes/../common/././algorithm.cpp \
- ../src/iterators/../shapes/../common/./././algorithm.hpp \
- ../src/iterators/../shapes/../common/./././../typing/types.hpp \
- ../src/iterators/../shapes/../common/./././../typing/./stl_wrapper/weak_ptr.hpp \
- ../src/iterators/../shapes/../common/./././_v_func.hpp \
- ../src/iterators/../shapes/../common/./../typing/allocator.cpp \
- ../src/iterators/../shapes/../common/./../typing/./allocator.hpp \
- ../src/iterators/../shapes/../common/./../typing/././types.hpp \
- ../src/iterators/../shapes/../common/./../random/generator.cpp \
- ../src/iterators/../shapes/../common/./../random/./generator.hpp \
- ../src/iterators/../shapes/../common/./../handlers/exception.hpp
+ ../src/iterators/../typing/utypes/types.hpp \
+ ../src/iterators/../typing/utypes/../gtypes/types.hpp \
+ ../src/iterators/../typing/utypes/../gtypes/./stl_wrapper/weak_ptr.hpp \
+ ../src/iterators/../typing/utypes/../utypes/mask.hpp \
+ ../src/iterators/../typing/utypes/../../common/vec1d.cpp \
+ ../src/iterators/../typing/utypes/../../common/./vec1d.hpp \
+ ../src/iterators/../typing/utypes/../../common/./../typing/gtypes/types.hpp \
+ ../src/iterators/../typing/utypes/../../common/./../handlers/exception.hpp \
+ ../src/iterators/../typing/utypes/../../common/./../typing/allocator.cpp \
+ ../src/iterators/../typing/utypes/../../common/./../typing/./allocator.hpp \
+ ../src/iterators/../typing/utypes/../../common/./../typing/././gtypes/types.hpp \
+ ../src/iterators/../typing/utypes/../../common/./../random/generator.cpp \
+ ../src/iterators/../typing/utypes/../../common/./../random/./generator.hpp \
+ ../src/iterators/../typing/utypes/../../common/././vec1d_ufunc.hpp \
+ ../src/iterators/../typing/utypes/../../common/././algorithm.cpp \
+ ../src/iterators/../typing/utypes/../../common/./././algorithm.hpp \
+ ../src/iterators/../typing/utypes/../../common/./././../typing/gtypes/types.hpp
 
 ../src/iterators/SequentialNdIterator.hpp:
 
-../src/iterators/../shapes/coords.hpp:
+../src/iterators/../typing/utypes/types.hpp:
 
-../src/iterators/../shapes/../common/vec1d.cpp:
+../src/iterators/../typing/utypes/../gtypes/types.hpp:
 
-../src/iterators/../shapes/../common/./vec1d.hpp:
+../src/iterators/../typing/utypes/../gtypes/./stl_wrapper/weak_ptr.hpp:
 
-../src/iterators/../shapes/../common/././algorithm.cpp:
+../src/iterators/../typing/utypes/../utypes/mask.hpp:
 
-../src/iterators/../shapes/../common/./././algorithm.hpp:
+../src/iterators/../typing/utypes/../../common/vec1d.cpp:
 
-../src/iterators/../shapes/../common/./././../typing/types.hpp:
+../src/iterators/../typing/utypes/../../common/./vec1d.hpp:
 
-../src/iterators/../shapes/../common/./././../typing/./stl_wrapper/weak_ptr.hpp:
+../src/iterators/../typing/utypes/../../common/./../typing/gtypes/types.hpp:
 
-../src/iterators/../shapes/../common/./././_v_func.hpp:
+../src/iterators/../typing/utypes/../../common/./../handlers/exception.hpp:
 
-../src/iterators/../shapes/../common/./../typing/allocator.cpp:
+../src/iterators/../typing/utypes/../../common/./../typing/allocator.cpp:
 
-../src/iterators/../shapes/../common/./../typing/./allocator.hpp:
+../src/iterators/../typing/utypes/../../common/./../typing/./allocator.hpp:
 
-../src/iterators/../shapes/../common/./../typing/././types.hpp:
+../src/iterators/../typing/utypes/../../common/./../typing/././gtypes/types.hpp:
 
-../src/iterators/../shapes/../common/./../random/generator.cpp:
+../src/iterators/../typing/utypes/../../common/./../random/generator.cpp:
 
-../src/iterators/../shapes/../common/./../random/./generator.hpp:
+../src/iterators/../typing/utypes/../../common/./../random/./generator.hpp:
 
-../src/iterators/../shapes/../common/./../handlers/exception.hpp:
+../src/iterators/../typing/utypes/../../common/././vec1d_ufunc.hpp:
+
+../src/iterators/../typing/utypes/../../common/././algorithm.cpp:
+
+../src/iterators/../typing/utypes/../../common/./././algorithm.hpp:
+
+../src/iterators/../typing/utypes/../../common/./././../typing/gtypes/types.hpp:
