+++ Release/src/multidim/matrix_stdout.d
@@ -1,52 +1,64 @@
 src/multidim/matrix_stdout.o: ../src/multidim/matrix_stdout.cpp \
  ../src/multidim/./matrix.hpp \
  ../src/multidim/./../iterators/RandomAccessNdIterator.hpp \
- ../src/multidim/./../iterators/./SequentialNdIterator.hpp \
- ../src/multidim/./../iterators/./../shapes/coords.hpp \
- ../src/multidim/./../iterators/./../shapes/../common/vec1d.cpp \
- ../src/multidim/./../iterators/./../shapes/../common/./vec1d.hpp \
- ../src/multidim/./../iterators/./../shapes/../common/././algorithm.cpp \
- ../src/multidim/./../iterators/./../shapes/../common/./././algorithm.hpp \
- ../src/multidim/./../iterators/./../shapes/../common/./././../typing/types.hpp \
- ../src/multidim/./../iterators/./../shapes/../common/./././../typing/./stl_wrapper/weak_ptr.hpp \
- ../src/multidim/./../iterators/./../shapes/../common/./././_v_func.hpp \
- ../src/multidim/./../iterators/./../shapes/../common/./../typing/allocator.cpp \
- ../src/multidim/./../iterators/./../shapes/../common/./../typing/./allocator.hpp \
- ../src/multidim/./../iterators/./../shapes/../common/./../typing/././types.hpp \
- ../src/multidim/./../iterators/./../shapes/../common/./../random/generator.cpp \
- ../src/multidim/./../iterators/./../shapes/../common/./../random/./generator.hpp \
- ../src/multidim/./../iterators/./../shapes/../common/./../handlers/exception.hpp
+ ../src/multidim/./../iterators/../typing/utypes/types.hpp \
+ ../src/multidim/./../iterators/../typing/utypes/../gtypes/types.hpp \
+ ../src/multidim/./../iterators/../typing/utypes/../gtypes/./stl_wrapper/weak_ptr.hpp \
+ ../src/multidim/./../iterators/../typing/utypes/../utypes/mask.hpp \
+ ../src/multidim/./../iterators/../typing/utypes/../../common/vec1d.cpp \
+ ../src/multidim/./../iterators/../typing/utypes/../../common/./vec1d.hpp \
+ ../src/multidim/./../iterators/../typing/utypes/../../common/./../typing/gtypes/types.hpp \
+ ../src/multidim/./../iterators/../typing/utypes/../../common/./../handlers/exception.hpp \
+ ../src/multidim/./../iterators/../typing/utypes/../../common/./../typing/allocator.cpp \
+ ../src/multidim/./../iterators/../typing/utypes/../../common/./../typing/./allocator.hpp \
+ ../src/multidim/./../iterators/../typing/utypes/../../common/./../typing/././gtypes/types.hpp \
+ ../src/multidim/./../iterators/../typing/utypes/../../common/./../random/generator.cpp \
+ ../src/multidim/./../iterators/../typing/utypes/../../common/./../random/./generator.hpp \
+ ../src/multidim/./../iterators/../typing/utypes/../../common/././vec1d_ufunc.hpp \
+ ../src/multidim/./../iterators/../typing/utypes/../../common/././algorithm.cpp \
+ ../src/multidim/./../iterators/../typing/utypes/../../common/./././algorithm.hpp \
+ ../src/multidim/./../iterators/../typing/utypes/../../common/./././../typing/gtypes/types.hpp \
+ ../src/multidim/./../iterators/../shapes/coords.hpp \
+ ../src/multidim/./../iterators/../shapes/../typing/utypes/types.hpp
 
 ../src/multidim/./matrix.hpp:
 
 ../src/multidim/./../iterators/RandomAccessNdIterator.hpp:
 
-../src/multidim/./../iterators/./SequentialNdIterator.hpp:
+../src/multidim/./../iterators/../typing/utypes/types.hpp:
 
-../src/multidim/./../iterators/./../shapes/coords.hpp:
+../src/multidim/./../iterators/../typing/utypes/../gtypes/types.hpp:
 
-../src/multidim/./../iterators/./../shapes/../common/vec1d.cpp:
+../src/multidim/./../iterators/../typing/utypes/../gtypes/./stl_wrapper/weak_ptr.hpp:
 
-../src/multidim/./../iterators/./../shapes/../common/./vec1d.hpp:
+../src/multidim/./../iterators/../typing/utypes/../utypes/mask.hpp:
 
-../src/multidim/./../iterators/./../shapes/../common/././algorithm.cpp:
+../src/multidim/./../iterators/../typing/utypes/../../common/vec1d.cpp:
 
-../src/multidim/./../iterators/./../shapes/../common/./././algorithm.hpp:
+../src/multidim/./../iterators/../typing/utypes/../../common/./vec1d.hpp:
 
-../src/multidim/./../iterators/./../shapes/../common/./././../typing/types.hpp:
+../src/multidim/./../iterators/../typing/utypes/../../common/./../typing/gtypes/types.hpp:
 
-../src/multidim/./../iterators/./../shapes/../common/./././../typing/./stl_wrapper/weak_ptr.hpp:
+../src/multidim/./../iterators/../typing/utypes/../../common/./../handlers/exception.hpp:
 
-../src/multidim/./../iterators/./../shapes/../common/./././_v_func.hpp:
+../src/multidim/./../iterators/../typing/utypes/../../common/./../typing/allocator.cpp:
 
-../src/multidim/./../iterators/./../shapes/../common/./../typing/allocator.cpp:
+../src/multidim/./../iterators/../typing/utypes/../../common/./../typing/./allocator.hpp:
 
-../src/multidim/./../iterators/./../shapes/../common/./../typing/./allocator.hpp:
+../src/multidim/./../iterators/../typing/utypes/../../common/./../typing/././gtypes/types.hpp:
 
-../src/multidim/./../iterators/./../shapes/../common/./../typing/././types.hpp:
+../src/multidim/./../iterators/../typing/utypes/../../common/./../random/generator.cpp:
 
-../src/multidim/./../iterators/./../shapes/../common/./../random/generator.cpp:
+../src/multidim/./../iterators/../typing/utypes/../../common/./../random/./generator.hpp:
 
-../src/multidim/./../iterators/./../shapes/../common/./../random/./generator.hpp:
+../src/multidim/./../iterators/../typing/utypes/../../common/././vec1d_ufunc.hpp:
 
-../src/multidim/./../iterators/./../shapes/../common/./../handlers/exception.hpp:
+../src/multidim/./../iterators/../typing/utypes/../../common/././algorithm.cpp:
+
+../src/multidim/./../iterators/../typing/utypes/../../common/./././algorithm.hpp:
+
+../src/multidim/./../iterators/../typing/utypes/../../common/./././../typing/gtypes/types.hpp:
+
+../src/multidim/./../iterators/../shapes/coords.hpp:
+
+../src/multidim/./../iterators/../shapes/../typing/utypes/types.hpp:
