+++ Release/src/typing/allocator.d
@@ -1,9 +1,9 @@
 src/typing/allocator.o: ../src/typing/allocator.cpp \
- ../src/typing/./allocator.hpp ../src/typing/././types.hpp \
- ../src/typing/./././stl_wrapper/weak_ptr.hpp
+ ../src/typing/./allocator.hpp ../src/typing/././gtypes/types.hpp \
+ ../src/typing/././gtypes/./stl_wrapper/weak_ptr.hpp
 
 ../src/typing/./allocator.hpp:
 
-../src/typing/././types.hpp:
+../src/typing/././gtypes/types.hpp:
 
-../src/typing/./././stl_wrapper/weak_ptr.hpp:
+../src/typing/././gtypes/./stl_wrapper/weak_ptr.hpp:
