+++ Release/tests/test_api/alloc_nd_matrix.d
@@ -4,21 +4,25 @@ tests/test_api/alloc_nd_matrix.o: ../tests/test_api/alloc_nd_matrix.cpp \
  ../tests/test_api/./../../src/multidim/./matrix_alloc.cpp \
  ../tests/test_api/./../../src/multidim/././matrix.hpp \
  ../tests/test_api/./../../src/multidim/././../iterators/RandomAccessNdIterator.hpp \
- ../tests/test_api/./../../src/multidim/././../iterators/./SequentialNdIterator.hpp \
- ../tests/test_api/./../../src/multidim/././../iterators/./../shapes/coords.hpp \
- ../tests/test_api/./../../src/multidim/././../iterators/./../shapes/../common/vec1d.cpp \
- ../tests/test_api/./../../src/multidim/././../iterators/./../shapes/../common/./vec1d.hpp \
- ../tests/test_api/./../../src/multidim/././../iterators/./../shapes/../common/././algorithm.cpp \
- ../tests/test_api/./../../src/multidim/././../iterators/./../shapes/../common/./././algorithm.hpp \
- ../tests/test_api/./../../src/multidim/././../iterators/./../shapes/../common/./././../typing/types.hpp \
- ../tests/test_api/./../../src/multidim/././../iterators/./../shapes/../common/./././../typing/./stl_wrapper/weak_ptr.hpp \
- ../tests/test_api/./../../src/multidim/././../iterators/./../shapes/../common/./././_v_func.hpp \
- ../tests/test_api/./../../src/multidim/././../iterators/./../shapes/../common/./../typing/allocator.cpp \
- ../tests/test_api/./../../src/multidim/././../iterators/./../shapes/../common/./../typing/./allocator.hpp \
- ../tests/test_api/./../../src/multidim/././../iterators/./../shapes/../common/./../typing/././types.hpp \
- ../tests/test_api/./../../src/multidim/././../iterators/./../shapes/../common/./../random/generator.cpp \
- ../tests/test_api/./../../src/multidim/././../iterators/./../shapes/../common/./../random/./generator.hpp \
- ../tests/test_api/./../../src/multidim/././../iterators/./../shapes/../common/./../handlers/exception.hpp \
+ ../tests/test_api/./../../src/multidim/././../iterators/../typing/utypes/types.hpp \
+ ../tests/test_api/./../../src/multidim/././../iterators/../typing/utypes/../gtypes/types.hpp \
+ ../tests/test_api/./../../src/multidim/././../iterators/../typing/utypes/../gtypes/./stl_wrapper/weak_ptr.hpp \
+ ../tests/test_api/./../../src/multidim/././../iterators/../typing/utypes/../utypes/mask.hpp \
+ ../tests/test_api/./../../src/multidim/././../iterators/../typing/utypes/../../common/vec1d.cpp \
+ ../tests/test_api/./../../src/multidim/././../iterators/../typing/utypes/../../common/./vec1d.hpp \
+ ../tests/test_api/./../../src/multidim/././../iterators/../typing/utypes/../../common/./../typing/gtypes/types.hpp \
+ ../tests/test_api/./../../src/multidim/././../iterators/../typing/utypes/../../common/./../handlers/exception.hpp \
+ ../tests/test_api/./../../src/multidim/././../iterators/../typing/utypes/../../common/./../typing/allocator.cpp \
+ ../tests/test_api/./../../src/multidim/././../iterators/../typing/utypes/../../common/./../typing/./allocator.hpp \
+ ../tests/test_api/./../../src/multidim/././../iterators/../typing/utypes/../../common/./../typing/././gtypes/types.hpp \
+ ../tests/test_api/./../../src/multidim/././../iterators/../typing/utypes/../../common/./../random/generator.cpp \
+ ../tests/test_api/./../../src/multidim/././../iterators/../typing/utypes/../../common/./../random/./generator.hpp \
+ ../tests/test_api/./../../src/multidim/././../iterators/../typing/utypes/../../common/././vec1d_ufunc.hpp \
+ ../tests/test_api/./../../src/multidim/././../iterators/../typing/utypes/../../common/././algorithm.cpp \
+ ../tests/test_api/./../../src/multidim/././../iterators/../typing/utypes/../../common/./././algorithm.hpp \
+ ../tests/test_api/./../../src/multidim/././../iterators/../typing/utypes/../../common/./././../typing/gtypes/types.hpp \
+ ../tests/test_api/./../../src/multidim/././../iterators/../shapes/coords.hpp \
+ ../tests/test_api/./../../src/multidim/././../iterators/../shapes/../typing/utypes/types.hpp \
  ../tests/test_api/./../../src/multidim/./matrix_ops.cpp \
  ../tests/test_api/./../../src/multidim/./matrix_stdout.cpp \
  ../tests/test_api/./../../src/linalg/linalg.cpp \
@@ -35,35 +39,43 @@ tests/test_api/alloc_nd_matrix.o: ../tests/test_api/alloc_nd_matrix.cpp \
 
 ../tests/test_api/./../../src/multidim/././../iterators/RandomAccessNdIterator.hpp:
 
-../tests/test_api/./../../src/multidim/././../iterators/./SequentialNdIterator.hpp:
+../tests/test_api/./../../src/multidim/././../iterators/../typing/utypes/types.hpp:
 
-../tests/test_api/./../../src/multidim/././../iterators/./../shapes/coords.hpp:
+../tests/test_api/./../../src/multidim/././../iterators/../typing/utypes/../gtypes/types.hpp:
 
-../tests/test_api/./../../src/multidim/././../iterators/./../shapes/../common/vec1d.cpp:
+../tests/test_api/./../../src/multidim/././../iterators/../typing/utypes/../gtypes/./stl_wrapper/weak_ptr.hpp:
 
-../tests/test_api/./../../src/multidim/././../iterators/./../shapes/../common/./vec1d.hpp:
+../tests/test_api/./../../src/multidim/././../iterators/../typing/utypes/../utypes/mask.hpp:
 
-../tests/test_api/./../../src/multidim/././../iterators/./../shapes/../common/././algorithm.cpp:
+../tests/test_api/./../../src/multidim/././../iterators/../typing/utypes/../../common/vec1d.cpp:
 
-../tests/test_api/./../../src/multidim/././../iterators/./../shapes/../common/./././algorithm.hpp:
+../tests/test_api/./../../src/multidim/././../iterators/../typing/utypes/../../common/./vec1d.hpp:
 
-../tests/test_api/./../../src/multidim/././../iterators/./../shapes/../common/./././../typing/types.hpp:
+../tests/test_api/./../../src/multidim/././../iterators/../typing/utypes/../../common/./../typing/gtypes/types.hpp:
 
-../tests/test_api/./../../src/multidim/././../iterators/./../shapes/../common/./././../typing/./stl_wrapper/weak_ptr.hpp:
+../tests/test_api/./../../src/multidim/././../iterators/../typing/utypes/../../common/./../handlers/exception.hpp:
 
-../tests/test_api/./../../src/multidim/././../iterators/./../shapes/../common/./././_v_func.hpp:
+../tests/test_api/./../../src/multidim/././../iterators/../typing/utypes/../../common/./../typing/allocator.cpp:
 
-../tests/test_api/./../../src/multidim/././../iterators/./../shapes/../common/./../typing/allocator.cpp:
+../tests/test_api/./../../src/multidim/././../iterators/../typing/utypes/../../common/./../typing/./allocator.hpp:
 
-../tests/test_api/./../../src/multidim/././../iterators/./../shapes/../common/./../typing/./allocator.hpp:
+../tests/test_api/./../../src/multidim/././../iterators/../typing/utypes/../../common/./../typing/././gtypes/types.hpp:
 
-../tests/test_api/./../../src/multidim/././../iterators/./../shapes/../common/./../typing/././types.hpp:
+../tests/test_api/./../../src/multidim/././../iterators/../typing/utypes/../../common/./../random/generator.cpp:
 
-../tests/test_api/./../../src/multidim/././../iterators/./../shapes/../common/./../random/generator.cpp:
+../tests/test_api/./../../src/multidim/././../iterators/../typing/utypes/../../common/./../random/./generator.hpp:
 
-../tests/test_api/./../../src/multidim/././../iterators/./../shapes/../common/./../random/./generator.hpp:
+../tests/test_api/./../../src/multidim/././../iterators/../typing/utypes/../../common/././vec1d_ufunc.hpp:
 
-../tests/test_api/./../../src/multidim/././../iterators/./../shapes/../common/./../handlers/exception.hpp:
+../tests/test_api/./../../src/multidim/././../iterators/../typing/utypes/../../common/././algorithm.cpp:
+
+../tests/test_api/./../../src/multidim/././../iterators/../typing/utypes/../../common/./././algorithm.hpp:
+
+../tests/test_api/./../../src/multidim/././../iterators/../typing/utypes/../../common/./././../typing/gtypes/types.hpp:
+
+../tests/test_api/./../../src/multidim/././../iterators/../shapes/coords.hpp:
+
+../tests/test_api/./../../src/multidim/././../iterators/../shapes/../typing/utypes/types.hpp:
 
 ../tests/test_api/./../../src/multidim/./matrix_ops.cpp:
