+++ src/common/algorithm.cpp
@@ -5,7 +5,6 @@
  */
 
 #include "./algorithm.hpp"
-#include "./_v_func.hpp"
 
 template<typename T>
 algorithm<T>::algorithm() = default;
