+++ src/common/algorithm.hpp
@@ -7,9 +7,9 @@
 #ifndef SRC_ALGORITHM_HPP
 #define SRC_ALGORITHM_HPP
 
-#include <queue>
+#include <algorithm>
 
-#include "../typing/types.hpp"
+#include "../typing/gtypes/types.hpp"
 
 template<typename T>
 class algorithm {
