+++ src/common/vec1d_ufunc.hpp
